+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Added: `unit-no-unknown` rule.
 - Fixed: `no-descending-specificity` no longer gets confused when the last part of a selector is a compound selector.
+- Fixed: regression causing `indentation` to complain about Sass maps.
 
 # 5.3.0